Fix not domain found for define command
The define action searchs for the domain name into the xml definition
to determine if the domain needs to be created or updated. The xml
variable contains the parsed definition but doesn't guarantee the
existence of the name tag. This change targets to fix the scenario
where the xml var is not empty but doesn't contain a name tag.