connection/libvirt_qemu: Mitigate a CPU hammering active wait loop
Once the `libvirt_qemu` connection plugin has spawned a command via the
qemu guest agent, it needs to wait for it to finish.
This is done in an active wait loop.

However, without any sleep, this wait loop consumes a lot of CPU.

On my machine, the following command:
```
ansible archlinux-vm -m command -a "sleep 10"
```
makes the `qemu-ga` of my VM consume 20% of CPU
and `systemd-journald` consume 7% of CPU because all qemu guest agent
commands are logged.

Adding a fixed value sleep time in each loop iteration could make tasks
execution run slower, especially if there are a lot of quick ones.

In order to not constantly consume a CPU for long-running tasks without
increasing the latency on quick ones, this change introduces an exponentially
increasing sleep in the active wait loop.
It guarantees that the latency increase it bounded up to 5% of the task
execution time while drastically reducing the CPU consumption.

Here are the impacts of this patch on the following command:
```
ansible archlinux-vm -m command -a "sleep 10"
```

| Measure                       |  Before  |  After   |
| :---------------------------- | -------: | -------: |
| Elapsed time                  | 11220 ms | 11389 ms |
| CPU time ansible host         |  3539 ms |  2655 ms |
| CPU libvirt VM ansible target |  3826 ms |   271 ms |

We can see that the CPU consumption has been drastically reduced without
increasing the elapsed time too much.