WIP: doc: clarify connection reqs for Windows guests
Windows guests are supported with the connection plugin, however in
order to trigger the required code path, the 'ansible_shell_type'
variable must be set to 'cmd' or 'powershell' for that host. Without
that, Ansible will default to `/bin/sh` which does not work.

This docs patch makes the requirements more clear.