Replace discouraged function listVolumes
According to the libvirt storage API reference, the function

- listVolumes (virStoragePoolListVolumes)

should not be used anymore: "The use of this
function is discouraged. Instead, use virStoragePoolListAllVolumes()."

Source:
https://libvirt.org/html/libvirt-libvirt-storage.html#virStoragePoolListVolumes

Compatibility / Risk:
The function "virStoragePoolListAllVolumes()" appeared in version
0.10.2 which was released in 2012. It seems rather unlikely that
somebody is still using an older unsupported libvirt version, so the
risk should be rather low.

Source:
https://libvirt.org/hvsupport.html#virStorageDriver