Replace functions listStoragePools and listDefinedStoragePools
According to the libvirt storage API reference, the functions

- listStoragePools (virConnectListStoragePools)
- listDefinedStoragePools (virConnectListDefinedStoragePools)

should not be used anymore: "The use of this function is discouraged.
Instead, use virConnectListAllStoragePools()" ... "This API solves the
race inherent between virConnectListStoragePools and
virConnectListDefinedStoragePools."

Source:
https://libvirt.org/html/libvirt-libvirt-storage.html#virConnectListAllStoragePools

Compatibility / Risk:
The function "virConnectListAllStoragePools()" appeared in version
0.10.2 which was released in 2012. It seems rather unlikely that
somebody is still using an older unsupported libvirt version, so the
risk should be rather low.

Source:
https://libvirt.org/hvsupport.html#virStorageDriver

The function can be used with or without a flag:
- get ALL pools with listAllStoragePools()
- get active pools with listAllStoragePools(2)
- get inactive pools with listAllStoragePools(1)

Before, loops were required to build a list of active and inactive
pools. Now, the function returns a list of all pools in just one call.

Co-authored-by: Andrew Klychkov <aaklychkov@mail.ru>