Fix virt module to undefine a domain with nvram or other metadata (#136)
Libvirt function undefine() is not able to delete nvram or other
metadata. Therefore it is replaced with undefineFlags() which is able to
handle it by using flags.

All possible flags are listed in 'ENTRY_UNDEFINE_FLAGS_MAP'. Integer 23
makes undefine successful in all cases (1 + 2 + 4 +
16).

Source:
https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainUndefineFlags

The flags nvram(4) and keep_nvram(8) are mutually exclusive.

To control the metadata handling during undefine, two new options
'force' and 'flags' are introduced including documentation and examples.

Compatibility / Risk:
The function undefineFlags() appeared in libvirt version 0.9.4 which was
release in 2011. It seems rather unlikely that somebody is still using
an older unsupported libvirt version. Additionally, if none of the new
module options are provided, then it behaves as before and maintains
backward compatibility, means the overall risk of this commit should be
rather low.

Source:
https://libvirt.org/hvsupport.html#virHypervisorDriver