fix up failing pylint test, use before assignment
The latest pylint test is failing with error that variable might be used
before it is assigned in a couple of try, except, finally blocks.

This patch defaults the variables to an empty string, as the variable
would be set to a string in either the try or the except.