fix Ansible temp directory path (fixes #161)
The libvirt_qemu connection module uses QEMU guest agent to copy files
and run commands on the remote host. Ansible uses the '~/' symbolic path
to the user's home directory for many commands and file paths.
Unfortunately, since QEMU guest agent doesn't run in a shell, it doesn't
resolve the '~/' path to the user's homedir.

Instead, the Ansible paths end up at '/~/' in the remote host. This
breaks some assumtions for some modules about the path to the Ansible
files. Some modules use the resolved homedir path and fail to find any
files in that location.

This fix first resolves the remote user homedir with a QEMU guest
command, caches the result, and then replaces '~/' found in any command
or remote file paths with the resolved absolute path to the user's
homedir.