Add virt_volume return key/value pairs for Type, Capacity and Allocation
Using 'virt_volume' command 'list_volumes' is returning a key 'info' which
contains a simple list of several numbers. For someone unfamiliar with the
libvirt internals the meaning of these values in 'info' are rather unclear.

This commit transforms the list of numbers and returns them in following
additional key/value pairs:

- 'Type': type of the volume as string
- 'Capacity': total size of the volume in bytes as integer
- 'Allocation': current allocated size of the volume in bytes as integer

Sources:
- https://libvirt.org/html/libvirt-libvirt-storage.html#virStorageVolInfo
- https://libvirt.org/html/libvirt-libvirt-storage.html#virStorageVolType

Create a new module_util to hold constants with the future goal to share code
more easily between modules.