Added ability to resize existing volumes
When volume already exists we need to be able to resize it. To preserve
idempotency existing volume capacity is compared with the definde one.
If size differs executes resize on the volume. We do not check if
allocated size is smaller than configured size, but libirt will throw an
error when resize to a lower volume will be attempted.

Calculation of the size from xml moved to a separate function to reduce
complexity of the create class method.
Volume resize moved to a more global place where it is attempted/tested
all the time.