Added virt_volume resize tests and description
Extended state description with the statement that it can be used to
resize existing volumes if new capacity is bigger than current.

Added "Tests for volume resize" to verify resize behaviour. It creates
volume of defined size, then check idempotency and finally attempts to
resize the volume. Assert checks resulting size of the volume.