Added ability to resize existing volumes (#226)
* Added ability to resize existing volumes

When volume already exists we need to be able to resize it. To preserve
idempotency existing volume capacity is compared with the definde one.
If size differs executes resize on the volume. We do not check if
allocated size is smaller than configured size, but libirt will throw an
error when resize to a lower volume will be attempted.

Calculation of the size from xml moved to a separate function to reduce
complexity of the create class method.
Volume resize moved to a more global place where it is attempted/tested
all the time.

* Added changelog fragment

* Added virt_volume resize tests and description

Extended state description with the statement that it can be used to
resize existing volumes if new capacity is bigger than current.

Added "Tests for volume resize" to verify resize behaviour. It creates
volume of defined size, then check idempotency and finally attempts to
resize the volume. Assert checks resulting size of the volume.

---------

Co-authored-by: Denys Mishchenko <denys.mishchenko@cleura.com>