[WIP] libvirt secret module and VirtModule common class
A bit of a new approach by using common base class VirtModule which
deduplicates shared functions like connection to a libvirt or ansible
module exit report text. Some additional diff functions which are usable
to show diff when the module run with the diff flag and potentially in a
check mode.

Added new module virt_secret which can be used to manage libvirt secrets